+++ tests/Fixtures/ValidatorSubclass.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace jbboehr\PhpstanLaravelValidation\Test\Fixtures;
+
+final class ValidatorSubclass extends \Illuminate\Validation\Validator
+{
+}